+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Regression tests for validating Scryfall bulkdata ([#278](https://github.com/arcavios/scooze/pull/278))
 - Add `scooze save cards` to the scooze CLI ([#287](https://github.com/arcavios/scooze/pull/287))
 - Add `spree` to supported `FrameEffect`s ([#293](https://github.com/arcavios/scooze/pull/293))
+- Add Templar Knight to unlimited copy list ([#326](https://github.com/arcavios/scooze/pull/326))
 
 #### Changed