+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 #### Added
 
 - Add support for hybrid colorless mana symbols ([#323](https://github.com/arcavios/scooze/pull/323))
-
+- Add support for pawprint (modal budget) symbol ([#327](https://github.com/arcavios/scooze/pull/327))
 
 ### [2.0.0] - 2024-05-06
 
@@ -18,6 +18,7 @@
 - Regression tests for validating Scryfall bulkdata ([#278](https://github.com/arcavios/scooze/pull/278))
 - Add `scooze save cards` to the scooze CLI ([#287](https://github.com/arcavios/scooze/pull/287))
 - Add `spree` to supported `FrameEffect`s ([#293](https://github.com/arcavios/scooze/pull/293))
+- Add Templar Knight to unlimited copy list ([#326](https://github.com/arcavios/scooze/pull/326))
 
 #### Changed