+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Add support for pawprint (modal budget) symbol ([#327](https://github.com/arcavios/scooze/pull/327))
 - Support for new symbols in BLB and MB2 ([#328](https://github.com/arcavios/scooze/pull/328))
 
+#### Docs
+
+- Deployment doc for releases and new sets ([#324](https://github.com/arcavios/scooze/pull/324))
+
 ### [2.0.0] - 2024-05-06
 
 [GitHub Release](https://github.com/arcavios/scooze/releases/tag/v2.0.0)